[FLINK-7811] Add Scala 2.12 profile
This changes some modules that had a _2.11 dependency but didn't expose
it in their module name to instead depend on the ${scala.binary.version}
dependency.

The main reason for this is to make the build self contained, before,
with the hard-dependency on 2.11, when buildig for 2.12 it would not be
clear where the dependency would come from because it is not created as
part of the build. This could lead to inconsistencies. For example, when
adding a new class in flink-runtime but not recompiling on 2.11 but only
on 2.12, the 2.12 tests would fail when using that new class because
they would use 2.11 dependencies that weren't rebuilt with the new
class.

We also don't build flink-scala-shell and flink-connector-kafka-0.8
because they don't work with Scala 2.12.

This also includes $PROFILE in dependency convergence check script. This
is in preparation for building with "-Dscala-212", where we hace to
exclude certain things.

We also exclude Kafka 0.8 from stages when building for Scala 2.12

And add Scala 2.12 to the release scripts