[FLINK-22748] Allow dynamic target topic selection in SQL Kafka sinks (#109)
[FLINK-22748][connector-kafka] Allow dynamic target topic selection in SQL Kafka sinks

Allows writing to different Kafka topics based on the topic metadata column value in SQL, and updates the Table API's KafkaDynamicSink to accept a List<String> topics instead of String topic as well as support topic-pattern. The list acts as an allow-list of acceptable values for the topic metadata column. topic-pattern for sinks is a pattern that must match topic metadata column values, or else an error is thrown.

If a single topic is provided, it is used by default for the target topic to produce to
If a list is provided, only that list of topics can be produced to
If a topic pattern is provided, it must match topic metadata column values