[FLINK-37611] Deflake ExactlyOnceKafkaWriterITCase#shouldAbortLingeringTransactions
The test was actually not working correctly since writer's get their unique prefix (pool rework PR). The test mostly succeeded since the partitions in which records were written was non-deterministic and more often than not the 3 records didn't meet in the 10 partitions which resulted in an incorrect pass.

Now the value is also passed as a key, which makes partitions assignment deterministic and for this test we just write the exact same value 3 times.